OR-1101 update source code structure
