+++ cmd/titond/main.go
@@ -1,10 +1,6 @@
 package main
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 
@@ -14,6 +10,7 @@ import (
 	"github.com/tokamak-network/tokamak-titond-backend/pkg/http"
 	"github.com/tokamak-network/tokamak-titond-backend/pkg/kubernetes"
 	"github.com/tokamak-network/tokamak-titond-backend/pkg/services"
+	apptest "github.com/tokamak-network/tokamak-titond-backend/test"
 	"github.com/urfave/cli/v2"
 )
 
@@ -27,8 +24,8 @@ func init() {
 		{
 			Name:    "check-swagger",
 			Aliases: []string{"g"},
-			Usage:   "Greet a person",
-			Action:  checkSwagger,
+			Usage:   "check swagger",
+			Action:  apptest.CheckSwagger,
 		},
 	}
 	app.Action = titond
@@ -92,62 +89,4 @@ func titond(ctx *cli.Context) error {
 	http.Run()
 
 	return nil
-}
-
-func checkSwagger(ctx *cli.Context) error {
-	fmt.Println("Check swagger")
-	http := http.NewHTTPServer(&http.Config{
-		Host: ctx.String("http.host"),
-		Port: ctx.String("http.port"),
-	}, nil)
-
-	numAPIs := len(http.R.Routes())
-
-	for _, route := range http.R.Routes() {
-		fmt.Printf("%-6s %-25s %s\n", route.Method, route.Path, route.Handler)
-	}
-	fmt.Println("Total APIs: ", numAPIs)
-	fmt.Println("Total APIs excluding swagger api: ", (numAPIs - 1))
-
-	jsonData, err := ioutil.ReadFile("./docs/swagger.json")
-	if err != nil {
-		fmt.Println("Read docs file:", err)
-		return err
-	}
-
-	var data map[string]interface{}
-	err = json.Unmarshal(jsonData, &data)
-	if err != nil {
-		fmt.Println("Docs data:", err)
-		return err
-	}
-	paths, exist := data["paths"]
-	if !exist {
-		fmt.Println("Failed to fetch api description in swagger file")
-		return errors.New("")
-	}
-	switch v := paths.(type) {
-	case map[string]interface{}:
-		numOfApisInDocs := countNumOfAPIsInDocs(v)
-		if numOfApisInDocs != numAPIs-1 {
-			return errors.New("missing swagger description for api")
-		}
-	default:
-		return errors.New("")
-	}
-	return nil
-}
-
-func countNumOfAPIsInDocs(paths map[string]interface{}) int {
-	counter := 0
-	for path, metadata := range paths {
-		fmt.Println("  [Path]:", path)
-		switch v := metadata.(type) {
-		case map[string]interface{}:
-			counter += len(v)
-		default:
-		}
-	}
-	fmt.Println("Api in docs ", counter)
-	return counter
 }
